Move head-block logic into separated class.
