+++ src/main/java/com/eternalcode/lobbyheads/head/HeadCommand.java
@@ -2,11 +2,7 @@ package com.eternalcode.lobbyheads.head;
 
 import com.eternalcode.lobbyheads.configuration.ConfigurationService;
 import com.eternalcode.lobbyheads.configuration.implementation.HeadsConfiguration;
-import com.eternalcode.lobbyheads.head.HeadInfo;
-import com.eternalcode.lobbyheads.head.HeadService;
 import com.eternalcode.lobbyheads.notification.NotificationAnnouncer;
-import org.bukkit.Material;
-import org.bukkit.block.Block;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -18,15 +14,15 @@ import java.util.List;
 public class HeadCommand implements CommandExecutor, TabCompleter {
 
     private final HeadsConfiguration headsConfiguration;
-    private final HeadService headService;
     private final ConfigurationService configurationService;
     private final NotificationAnnouncer notificationAnnouncer;
+    private final HeadBlockService headBlockService;
 
-    public HeadCommand(HeadsConfiguration headsConfiguration, HeadService headService, ConfigurationService configurationService, NotificationAnnouncer notificationAnnouncer) {
+    public HeadCommand(HeadsConfiguration headsConfiguration, ConfigurationService configurationService, NotificationAnnouncer notificationAnnouncer, HeadBlockService headBlockService) {
         this.headsConfiguration = headsConfiguration;
-        this.headService = headService;
         this.configurationService = configurationService;
         this.notificationAnnouncer = notificationAnnouncer;
+        this.headBlockService = headBlockService;
     }
 
     @Override
@@ -51,8 +47,8 @@ public class HeadCommand implements CommandExecutor, TabCompleter {
         String subcommand = args[0].toLowerCase();
 
         switch (subcommand) {
-            case "add" -> this.createHead(player);
-            case "remove" -> this.removeHead(player);
+            case "add" -> this.headBlockService.createHead(player);
+            case "remove" -> this.headBlockService.removeHead(player);
             case "reload" -> {
                 this.configurationService.reload();
                 this.notificationAnnouncer.sendMessage(player, this.headsConfiguration.messages.configurationReloaded);
@@ -62,41 +58,6 @@ public class HeadCommand implements CommandExecutor, TabCompleter {
         return true;
     }
 
-    void removeHead(Player player) {
-        Block block = player.getTargetBlock(null, 5);
-
-        for (HeadInfo headInfo : this.headsConfiguration.heads) {
-            if (headInfo.getLocation().equals(block.getLocation())) {
-                this.headService.removeHologram(headInfo);
-                this.headsConfiguration.heads.remove(headInfo);
-                this.headsConfiguration.save();
-                this.notificationAnnouncer.sendMessage(player, this.headsConfiguration.messages.headRemoved);
-
-                return;
-            }
-        }
-    }
-
-    void createHead(Player player) {
-        Block block = player.getTargetBlock(null, 5);
-
-        if (block.getType() != Material.PLAYER_HEAD) {
-            this.notificationAnnouncer.sendMessage(player, this.headsConfiguration.messages.youAreNotLookingAtHead);
-            return;
-        }
-
-        if (this.headService.find(block.getLocation()).isPresent()) {
-            this.notificationAnnouncer.sendMessage(player, this.headsConfiguration.messages.headAlreadyExists);
-            return;
-        }
-
-        HeadInfo headInfo = new HeadInfo(block.getLocation(), player.getName(), player.getUniqueId());
-
-        this.headService.createHologram(player, headInfo, this.headsConfiguration.head.headFormat);
-        this.headsConfiguration.heads.add(headInfo);
-        this.headsConfiguration.save();
-        this.notificationAnnouncer.sendMessage(player, this.headsConfiguration.messages.headAdded);
-    }
 
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String label, String[] args) {
